+++ src/common/routes/components.js
@@ -110,5 +110,5 @@ export default {
   WebsiteRestrictions: loableComponent(
     import('../../client/websites/WebsiteTools/Restrictions/WebsiteRestrictions'),
   ),
-  MainPage: loableComponent(import('../../client/components/MainPage/MainPage')),
+  MainBody: loableComponent(import('../../client/components/LayoutComponents/MainBody')),
 };
