+++ src/common/routes/routes.js
@@ -8,37 +8,6 @@ const routes = {
   component: Wrapper,
   routes: [
     {
-      path: `/:sortBy(${URL.FEED.tabs})?/:category?`,
-      component: Views.MainPage,
-      routes: [
-        {
-          path: '/confirmation',
-          exact: true,
-          component: Views.ConfirmationModal,
-        },
-        {
-          path: '/notifications-list',
-          component: Views.Notifications,
-        },
-        {
-          path: '/rewards-list',
-          component: Views.RewardsList,
-        },
-        {
-          path: '/feed/:name',
-          component: Views.ObjectFeed,
-        },
-        {
-          path: '/blog/@:name',
-          component: Views.UserProfile,
-        },
-        {
-          path: '/:sortBy(trending|created|hot)?/:category?',
-          component: Views.SubFeed,
-        },
-      ],
-    },
-    {
       path: '/bookmarks',
       exact: true,
       component: Views.Bookmarks,
@@ -430,39 +399,37 @@ const routes = {
       path: '/exit',
       component: Views.ExitPage,
     },
-    // {
-    //   path: `/:sortBy(${URL.FEED.tabs})?/:category?`,
-    //   component: Page,
-    //   exact: true,
-    //   pathScope: '',
-    //   routes: [
-    //     {
-    //       path: '/confirmation',
-    //       exact: true,
-    //       component: Views.ConfirmationModal,
-    //     },
-    //     {
-    //       path: '/notifications-list',
-    //       component: Views.Notifications,
-    //     },
-    //     {
-    //       path: '/rewards-list',
-    //       component: Views.RewardsList,
-    //     },
-    //     {
-    //       path: '/feed/:name',
-    //       component: Views.ObjectFeed,
-    //     },
-    //     {
-    //       path: '/blog/@:name',
-    //       component: Views.UserProfile,
-    //     },
-    //     {
-    //       path: '/:sortBy(trending|created|hot)?/:category?',
-    //       component: Views.SubFeed,
-    //     },
-    //   ],
-    // },
+    {
+      path: `/:sortBy(${URL.FEED.tabs})?/:category?`,
+      component: Views.MainBody,
+      routes: [
+        {
+          path: '/confirmation',
+          exact: true,
+          component: Views.ConfirmationModal,
+        },
+        {
+          path: '/notifications-list',
+          component: Views.Notifications,
+        },
+        {
+          path: '/rewards-list',
+          component: Views.RewardsList,
+        },
+        {
+          path: '/feed/:name',
+          component: Views.ObjectFeed,
+        },
+        {
+          path: '/blog/@:name',
+          component: Views.UserProfile,
+        },
+        {
+          path: '/:sortBy(trending|created|hot)?/:category?',
+          component: Views.SubFeed,
+        },
+      ],
+    },
     {
       path: '*',
       component: Views.Error404,
